+++ logic_pkg/main_menu_logic.py
@@ -1,5 +1,4 @@
-from ui_controller import ui_controller
-from logic_controller import logic_controller
+from controllers import ui_controller, logic_controller
 import requests
 
 def adminEventLoop(window, event, values):
