+++ logic_pkg/redeem_booking_logic.py
@@ -1,6 +1,5 @@
 import PySimpleGUI as sg
-from ui_controller import ui_controller
-from logic_controller import logic_controller
+from controllers import ui_controller, logic_controller
 from utils import utils
 
 def redeemBookingLoop(window, event, values):
